Multi now working, read desc
Unfortunately the jvm tries to do much for you as always and slows stuff down. While there is an increase in speed it is at best a quarter. Jvm thinks that rather than just reading from the caches it must treat shared vars as volatile, and actually go searching for them. Which would be a problem if the code didnt take into account that issue already and deal with it in an efficient way.